Refactored code a bit
