+++ src/main/java/dev/symphony/harmony/mixin/transportation/riptide/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains multiple features related to Trident's Riptide enchantment
+ */
+package dev.symphony.harmony.mixin.transportation.riptide;
